tests/requirements.txt
pytest
kafka-python
cassandra-driver
requests
python-dotenv
